refactor(mcp): extract params / job-runtime / pagination
Structural move: mcp/mod.rs (2024 lines) -> 1448, by extracting the
macro-free pieces into siblings —
- params.rs       the *Params tool-input structs + serde defaults
- job_runtime.rs  current-thread runtime bridge, sync error marking, JobSnapshot
- pagination.rs   result pagination + untrusted-content provenance banner

The rmcp-macro-annotated blocks (`#[tool_router] impl DalfoxMcp` with the six
`#[tool]` handlers, `impl Default`, the lifecycle impl, `impl ServerHandler`,
run_mcp_server) stay in mod.rs — a proc-macro impl must remain one block.
The tool-input structs stay pub(crate) and are re-exported so
`crate::mcp::ScanWithDalfoxParams` (the job REST/MCP parity test) resolves
unchanged; helpers are pub(super) reached via `use super::*`; tests.rs
untouched. 82 mcp + 38 job tests pass, clippy + rustfmt clean.
